Add a WPT test to verify the mouse click count in infrastructure/
We should have a WPT test to verify the mouse click count, such as
double click and triple click. For the click count is greater than 3, we
should increase it when the OS is Windows and Mac, reset to 1 otherwise.
Also reset the mouse click count, if the mouse moves to a different
locations.

Bug: 1119679
Change-Id: I755c1f49a470598391440802e2ad62817a958777