FontAccess: Chrome UI for a chooser
This CL implements a Chrome UI that allows users to pick one or more
fonts, resulting from a request originating from a Blink API call.

This change implements a multi-select chooser that offers the option to
"import" fonts to sites as well as the ability to cancel. It is
implemented as a permission bubble initially.

Fixed: 1138621
Change-Id: I9f4a0bd7a990823d2de0c19372cbedc4e811d7d6