[LayoutNG] Fieldset borders
After passing GetBackgroundBleedAvoidance() when painting the fieldset
border, there were cases when the fieldset border was not painted
correctly (when a border-radius was set as well as a border-color
with an alpha < 1). To fix this, we need to make further
adjustments based on the border when BleedAvoidanceIsClipping() is
true, similar to how other paint algorithms handle this case.

Bug: 1150372
Change-Id: Ia191f1348a449d49f9338b1cb7b9b3a5a023a5f2